try 3 groups
