+++ miss_hit.cfg
@@ -38,5 +38,5 @@ tab_width: 2
 # metrics limit for the code quality (https://florianschanda.github.io/miss_hit/metrics.html)
 metric "cnest": limit 5
 metric "file_length": limit 750
-metric "cyc": limit 19
+metric "cyc": limit 20
 metric "parameters": limit 7
