+++ tests/data/3_groups/models/model-vismotion_smdl.json
@@ -71,7 +71,7 @@
       "Name": "dataset_level",
       "GroupBy": [
         "contrast",
-        "Group"
+        "diagnostic"
       ],
       "Model": {
         "Type": "glm",
@@ -86,7 +86,7 @@
     {
       "Level": "Dataset",
       "Name": "between_groups",
-      "Description": "two sample t-test",
+      "Description": "one way anova",
       "GroupBy": [
         "contrast"
       ],
@@ -94,16 +94,16 @@
         "Type": "glm",
         "X": [
           1,
-          "group"
+          "diagnostic"
         ]
       },
       "Contrasts": [
         {
           "Name": "VisMot_gt_VisStat",
           "ConditionList": [
-            "group.ctrl",
-            "group.blind",
-            "group.relative"
+            "diagnostic.ctrl",
+            "diagnostic.blind",
+            "diagnostic.relative"
           ],
           "Weights": [
             1,
